tests/wpt/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-percent-height.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<table cellspacing="0" cellpadding="0" style="width:100px; height:60px;">
  <tr>
    <td>
      <fieldset style="border:none; padding:0; height:100%; margin:0; margin-top:13px;">
        <div><div id="elm"></div></div>
      </fieldset>
    </td>
  </tr>
</table>
<script>
// crbug.com/1138204. Though the specification doesn't mention this behavior,
// there must be no doubt about the expected behavior.
test(() => {
  const fieldset = document.querySelector('fieldset');
  const initialHeight = fieldset.offsetHeight;
  document.querySelector('#elm').style.display = 'none';
  assert_equals(fieldset.offsetHeight, initialHeight);
}, 'Fieldset with a percent height should not increase the height on every reflow.');
</script>
